[Enh](Layers): Added more modularity into config.
