+++ lua/config/Workspace/true_zen/init.lua
@@ -1,0 +1,4 @@
+return {
+  settings = require('config.Workspace.true_zen.settings'),
+  keymappings = require('config.Workspace.true_zen.keymappings'),
+}
